+++ internal/parse/testdata/watch.go
@@ -1,0 +1,18 @@
+//go:build stave
+
+package main
+
+import (
+	"github.com/yaklabco/stave/pkg/watch"
+)
+
+func WatchTarget() {
+	watch.Watch("*.go")
+}
+
+func NonWatchTarget() {
+}
+
+func WatchDepsTarget() {
+	watch.Deps(NonWatchTarget)
+}